move account list to it's own component and share
