+++ src/components/dialogs/SwitchAccount.tsx
@@ -7,12 +7,9 @@ import {useAccountSwitcher} from '#/lib/hooks/useAccountSwitcher'
 import {type SessionAccount, useSession} from '#/state/session'
 import {useLoggedOutViewControls} from '#/state/shell/logged-out'
 import {useCloseAllActiveElements} from '#/state/util'
-import {AccountItem} from '#/screens/Login/ChooseAccountForm'
-import {atoms as a, useTheme} from '#/alf'
+import {atoms as a} from '#/alf'
 import * as Dialog from '#/components/Dialog'
-import {ChevronRight_Stroke2_Corner0_Rounded as Chevron} from '#/components/icons/Chevron'
-import {Loader} from '#/components/Loader'
-import {Button} from '../Button'
+import {AccountList} from '../AccountList'
 import {Text} from '../Typography'
 
 export function SwitchAccountDialog({
@@ -20,9 +17,8 @@ export function SwitchAccountDialog({
 }: {
   control: Dialog.DialogControlProps
 }) {
-  const t = useTheme()
   const {_} = useLingui()
-  const {isSwitchingAccounts, currentAccount, accounts} = useSession()
+  const {currentAccount} = useSession()
   const {onPressSwitchAccount} = useAccountSwitcher()
   const {setShowLoggedOut} = useLoggedOutViewControls()
   const closeAllActiveElements = useCloseAllActiveElements()
@@ -54,58 +50,11 @@ export function SwitchAccountDialog({
           </Text>
         </View>
 
-        {isSwitchingAccounts ? (
-          <Loader size="xl" />
-        ) : (
-          <View
-            style={[
-              a.rounded_md,
-              a.overflow_hidden,
-              a.border,
-              t.atoms.border_contrast_low,
-            ]}>
-            {accounts.map(account => (
-              <React.Fragment key={account.did}>
-                <AccountItem
-                  account={account}
-                  onSelect={onSelectAccount}
-                  isCurrentAccount={account.did === currentAccount?.did}
-                />
-                <View style={[a.border_b, t.atoms.border_contrast_low]} />
-              </React.Fragment>
-            ))}
-            <Button
-              testID="chooseAddAccountBtn"
-              style={[a.flex_1]}
-              onPress={onPressAddAccount}
-              label={_(msg`Login to account that is not listed`)}>
-              {({hovered, pressed}) => (
-                <View
-                  style={[
-                    a.flex_1,
-                    a.flex_row,
-                    a.align_center,
-                    {height: 48},
-                    (hovered || pressed) && t.atoms.bg_contrast_25,
-                  ]}>
-                  <Text
-                    style={[
-                      a.align_baseline,
-                      a.flex_1,
-                      a.flex_row,
-                      a.py_sm,
-                      {paddingLeft: 48},
-                    ]}>
-                    <Trans>Add account</Trans>
-                  </Text>
-                  <Chevron size="sm" style={[t.atoms.text, a.mr_md]} />
-                </View>
-              )}
-            </Button>
-          </View>
-        )}
-
-        <Dialog.Close />
+        <AccountList
+          onSelectAccount={onSelectAccount}
+          onSelectOther={onPressAddAccount}
+          otherLabel={_(msg`Add account`)}
+        />
       </Dialog.ScrollableInner>
     </Dialog.Outer>
   )
